+++ safe-coloured-text-terminfo/src/Text/Colour/Term.hs
@@ -18,11 +18,6 @@ putChunksUtf8 = hPutChunksUtf8 stdout
 putChunksLocale :: [Chunk] -> IO ()
 putChunksLocale = hPutChunksLocale stdout
 
--- | Deprecated synonym of putChunksUtf8
-putChunks :: [Chunk] -> IO ()
-putChunks = putChunksUtf8
-{-# DEPRECATED putChunks "Use putChunksLocale, or putChunksUtf8 if you must." #-}
-
 -- | Print a list of chunks to the given 'Handle'
 --
 -- This function will use 'getTerminalCapabilitiesHandle' on the given handle.
@@ -40,8 +35,3 @@ hPutChunksLocale :: Handle -> [Chunk] -> IO ()
 hPutChunksLocale h cs = do
   tc <- getTerminalCapabilitiesFromHandle h
   hPutChunksLocaleWith tc h cs
-
--- | Deprecated synonym of 'hPutChunksUtf8'
-hPutChunks :: Handle -> [Chunk] -> IO ()
-hPutChunks = hPutChunksUtf8
-{-# DEPRECATED hPutChunks "Use hPutChunksLocale, or hPutChunksUtf8 if you must." #-}